ipc: move ipc host ptable ops to dedicated unit
Ipc private data needed by ptable processing removed from
platforms that do not use it.

Less ifdef-s in the main common ipc code, another new unit
compiled for platforms that require ptables.

Dependency on the private data removed from the main ipc code
and moved to the platform specific code. A dependency between
ipc-host-ptable and platform's ipc private data is easier to
maintain then duplicated code in the main ipc units.
