+++ src/ipc/CMakeLists.txt
@@ -14,3 +14,8 @@ if (CONFIG_TRACE)
 	add_local_sources(sof
 		dma-copy.c)
 endif()
+
+if (CONFIG_HOST_PTABLE)
+	add_local_sources(sof
+		ipc-host-ptable.c)
+endif()
